Remove the lint marker from more fix tests
Some of the tests were fairly seriously broken, so this fixes them.

Some of the tests were testing that we don't provide a fix when there is
no lint, but I don't think that it's useful to test that.

One test required no changes other than removing the marker.

Change-Id: I6d85fd98506fa20d8bf4996d30603feb8b1ff252
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/134500
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>