completion driver wip
Gotta start somewhere.

All subject to revision.  (Names especially.)

This gets us an end-to-end test and a starting point for client algorithm implementations.

Note that this was designed to run outside `test/`.  We should be able to move it out if so desired.

Needless to say any thoughts on any of it are welcome!

Change-Id: Iecf5c7ce854f619aa5677454091ceb68dd45827b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/134542
Commit-Queue: Phil Quitslund <pquitslund@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>