[vm/nnbd] Fix RangeError in NoSuchMethodError._toStringDeprecated
NoSuchMethodError._invocationType is actually always -1 which is out of
range. Before NNBD migration we had an explicit range check which helped
to avoid throwing exception, but it was replaced with assert during
migration. This CL restores the range check.

Fixes co19_2/LibTest/core/NoSuchMethodError/toString_A01_t01 with
unforked SDK.

Change-Id: I91abe92491fe3a15aa422a41fefe6538baac5621
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/134875
Reviewed-by: Siva Annamalai <asiva@google.com>