Pass a "--default-language-version" flag to analyzer for legacy tests.
When the test runner runs a legacy test with the non-nullable experiment
enabled, this passes an extra command line flag to tell analyzer to
treat the test library as opted out of NNBD.

This CL also removes the guard that made the test runner skip all legacy
tests when running analyzer with the non-nullable experiment enabled.
We want to run those tests now and (when analyzer supports the new
command line flag) they should be correctly treated as opted out.

This change also removes the tests package config file since that no
longer needs to be used to set the language version. It also removes the
code to generate a package config for the multitests. This CL replaces
both of those mechanisms.

Change-Id: Iedc1e6111bc0623ac21263499f3f5d11bd6fdf7c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/135346
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Bob Nystrom <rnystrom@google.com>