Gather more code completion metrics
The code is now gathering both element and token data in several places
where it used to gather only one of the two (using a new utility method).

The metrics are now better labeled to make the output easier to read.

It's also gathering a new metric: the distance between the static type
of a target and the class in which a referenced member is declared.

Change-Id: I566c02b069749221983ec960b8a67b293052bd93
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/135400
Reviewed-by: Jaime Wren <jwren@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>