[nnbd_migration] Fix stack overflow on missing element decorated type.
This was causing 9k exceptions and slowing the flutter migration from
2:40 to 14:40.

There are 52 elements that don't get a decorated type in flutter due to
the abnormal usage of part files for the animated icons test.

The re-resolution stage of fixbuilder would then try to get the
finalType of these elements, but no decorated type for them exists, so
it would throw this exception. However, since the throw expression
stringifys the element, the element would go back to the hooks that try
to turn it into a finalType, which results in the same failure, until a
stack overflow.

Change-Id: Ib4c4746728e8bc6e583e133e91dcc0ba83e1b911
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/136040
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Mike Fairhurst <mfairhurst@google.com>