Migration: do not allow RHS of a typedef to be marked as nullable.
We should consider the nullability of the right hand side of a typedef
to be "never", since a `?` is not allowed on there.  Instead, we
should allow each use of a typedef to have its own independent
nullability.

Fixes a bug wherein we would erroneously migrate code like this:

  typedef F<T> = int Function(T);
  F<String> f = null;

Into the following:

  typedef F<T> = int Function(T)?;
  F<String>? f = null;

(Note the bogus `?` after `int Function(T)`).

Change-Id: I8535645303fc644c73617684b8bb3488462589ee
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/136900
Reviewed-by: Mike Fairhurst <mfairhurst@google.com>