Migration: create object model for edit details.
We still exchange information between the client and server in JSON
format, but on the server side, we build the object and then serialize
it (rather than build the JSON directly); conversely, on the client
side we deserialize the JSON back into objects and then use them.
This makes the data interchange format easier to understand because it
is captured in class declarations.

Change-Id: Ia20293ce2713aa1793baf6e7ca8141da5615ab4a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/138011
Reviewed-by: Samuel Rawlins <srawlins@google.com>