[nnbd_migration] Apply the changes to disk on apply-migration request
This integrates fairly well with the new dartdev ux. We may be able to
improve the wording, etc., but given the fact that it has to be rerun
with --apply-edits, there is no need to notify anyone that the edits
occurred (currently). I'll take another look at potential wording
improvements in a follow-up CL.

In the dartfix ux, the dartfix driver doesn't get notified of the
changes and will ask if you want to apply, and this is not as good :) If
you say yes, you will apply the edits twice. Given that there is already
a CL to delete that code, I didn't change this in dartfix and don't plan
to.

Change-Id: Ie822a5761d05b3dd08b5a793f812a331d4657598
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/138203
Reviewed-by: Paul Berry <paulberry@google.com>