[dart:io] drain buffer within a single Read
Previous change on _NativeSocket::Read has a consequence of increasing the number of Read().
This change is to make Read() to drain as many bytes as possible to prevent Read event being posted again. But the side effect is that every single read() will become slower.

Bug: https://b.corp.google.com/issues/150902962
Change-Id: I58cc1e39ae598e2f1584da22d07ad1db3106bd39
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/138909
Commit-Queue: Zichang Guo <zichangguo@google.com>
Reviewed-by: Siva Annamalai <asiva@google.com>